Move code only used by custom_widgets out of custom_tools
